OSIS-5896 renamed functions
