+++ ddd/logic/learning_unit/use_case/write/create_effective_class_service.py
@@ -28,8 +28,8 @@ from django.db import transaction
 from ddd.logic.learning_unit.builder.learning_unit_identity_builder import LearningUnitIdentityBuilder
 from ddd.logic.learning_unit.commands import CreateEffectiveClassCommand
 from ddd.logic.learning_unit.domain.model.effective_class import EffectiveClassIdentity
-from ddd.logic.learning_unit.domain.service.can_access_creation_effective_class import CanCreateEffectiveClass
-from ddd.logic.learning_unit.domain.service.create_effective_class import SaveEffectiveClass
+from ddd.logic.learning_unit.domain.service.can_save_effective_class import CanCreateEffectiveClass
+from ddd.logic.learning_unit.domain.service.save_effective_class import SaveEffectiveClass
 from ddd.logic.learning_unit.repository.i_effective_class import IEffectiveClassRepository
 from ddd.logic.learning_unit.repository.i_learning_unit import ILearningUnitRepository
 
@@ -45,7 +45,7 @@ def create_effective_class(
     learning_unit = learning_unit_repository.get(
         entity_id=LearningUnitIdentityBuilder.build_from_code_and_year(cmd.learning_unit_code, cmd.year)
     )
-    CanCreateEffectiveClass().raise_if_can_not_access(
+    CanCreateEffectiveClass().verify(
         learning_unit=learning_unit,
         learning_unit_repository=learning_unit_repository
     )
