+++ ddd/logic/learning_unit/use_case/write/update_effective_class_service.py
@@ -27,8 +27,8 @@ from ddd.logic.learning_unit.builder.effective_class_identity_builder import Eff
 from ddd.logic.learning_unit.builder.learning_unit_identity_builder import LearningUnitIdentityBuilder
 from ddd.logic.learning_unit.commands import UpdateEffectiveClassCommand
 from ddd.logic.learning_unit.domain.model.effective_class import EffectiveClassIdentity
-from ddd.logic.learning_unit.domain.service.can_access_creation_effective_class import CanCreateEffectiveClass
-from ddd.logic.learning_unit.domain.service.create_effective_class import SaveEffectiveClass
+from ddd.logic.learning_unit.domain.service.can_save_effective_class import CanCreateEffectiveClass
+from ddd.logic.learning_unit.domain.service.save_effective_class import SaveEffectiveClass
 from ddd.logic.learning_unit.repository.i_effective_class import IEffectiveClassRepository
 from ddd.logic.learning_unit.repository.i_learning_unit import ILearningUnitRepository
 
@@ -43,7 +43,7 @@ def update_effective_class(
     learning_unit = learning_unit_repository.get(
         entity_id=LearningUnitIdentityBuilder.build_from_code_and_year(cmd.learning_unit_code, cmd.year)
     )
-    CanCreateEffectiveClass().raise_if_can_not_access(
+    CanCreateEffectiveClass().verify(
         learning_unit=learning_unit,
         learning_unit_repository=learning_unit_repository
     )
